[refactoring] step2: Mine 뷰로직 도메인에서 분리 및 테스트 실패 수정
