+++ src/main/kotlin/minesweeper/domain/mine/MineShape.kt
@@ -1,0 +1,5 @@
+package minesweeper.domain.mine
+
+enum class MineShape(val shape: String) {
+    MINE("*"), SHAPE("C")
+}
