+++ src/main/kotlin/minesweeper/main.kt
@@ -1,6 +1,7 @@
 package minesweeper
 
 import minesweeper.domain.board.size.MineSweeperBoardSize
+import minesweeper.domain.mine.Mine
 import minesweeper.ui.InputView
 import minesweeper.ui.OutputView
 
@@ -11,9 +12,10 @@ fun main() {
     val width = InputView.inputMineSweeperGameValue()
     OutputView.enterMineCount()
     val mineCount = InputView.inputMineSweeperGameValue()
+    val mine = Mine(mineCount)
     val mineSweeperBoardSize = MineSweeperBoardSize(width = width, height = height)
 
-    val mineSweeper = MineSweeper(board = mineSweeperBoardSize, mineCount = mineCount)
+    val mineSweeper = MineSweeper(board = mineSweeperBoardSize, mine = mine)
     OutputView.startMineSweeper()
 
     val mineSweeperBoard = mineSweeper.createBoard()
